Add settings page
A page for editing user settings was added.
